simplesocial/requirements.txt
alabaster==0.7.12
asgiref==3.5.0
Babel==2.9.1
beautifulsoup4==4.11.1
bootstrap-py==1.0.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Cython==0.29.28
defusedxml==0.7.1
Django==3.2.13
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-bootstrap3==21.2
django-braces==1.15.0
djangorestframework==3.13.1
docutils==0.17.1
Faker==3.0.1
gitdb2==2.0.6
GitPython==2.1.15
idna==2.10
imagesize==1.3.0
importlib-metadata==2.1.3
ipaddress==1.0.23
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pguard==0.4.0
Pillow==9.1.0
pip==22.1
pycparser==2.21
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
setuptools==61.2.0
six==1.16.0
smmap==3.0.5
smmap2==3.0.1
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==1.8.6
sphinxcontrib-websupport==1.1.2
sqlparse==0.4.1
text-unidecode==1.3
typing==3.10.0.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wheel==0.35.1
wincertstore==0.2
zipp==3.8.0
